Implement CSV export for saved spectra
Exported spectra always use the currently selected reference, not
necessarily the one they were created with. Not sure if this is
the ideal behavior.

Changing calibration between reference and exported is also an
issue perhaps worth addressing, but not specific to saving as it
also affects display of transmittance/absorbance.

One module was split up to avoid a circular dependency on the store
that prevented all reducers from loading.

A bug in the transmittance calculation was also fixed that made it
equivalent to a raw scale, as the last reference y value was always
selected. A typo caused num < Math.abs(undefined - num) which is
always false due to NaN.