Use pyinstaller in a more cross-platform manner
Spec files are platform-specific. This commit moves us away from
them in favor of the CLI. It also adds some platform-specific
arguments to create more seamless results, and moves us to onefile
bundles.

Replacing the Makefile and script with a Python script will yield
a truly cross-platform build, paving the way for CI/CD in the
future.